pyproject.toml
[project]
name = "self-evolving-agents"
version = "0.1.0"
description = "Self-evolving AI agents that learn from interaction patterns"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Core Dependencies
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
    # Synthetic Data Generation
    "openai>=1.0.0",
    "google-genai>=0.1.0",
    # Demo & Development
    "streamlit>=1.29.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    # Data Processing
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    # CLI Experience
    "rich>=13.7.0",
    "click>=8.1.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
# ML dependencies for local fine-tuning (optional - most users use Anthropic API)
ml = [
    "transformers>=4.36.0",
    "peft>=0.7.0",
    "datasets>=2.16.0",
    "torch>=2.1.0,<2.5.0",  # Pin to known-good versions
]

# Development tools
dev = [
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
